Remove duplicate bit_reverse
