+++ src/core/poly/utils.rs
@@ -53,28 +53,9 @@ pub(super) fn repeat_value<T: Copy>(values: &[T], duplicity: usize) -> Vec<T> {
     res
 }
 
-/// Bit reverses a slice.
-///
-/// # Panics
-///
-/// Panics if the length of the slice is not a power of two.
-pub(super) fn bit_reverse<T, U: AsMut<[T]>>(mut v: U) -> U {
-    let n = v.as_mut().len();
-    assert!(n.is_power_of_two());
-    let log_n = n.ilog2();
-    for i in 0..n {
-        let j = i.reverse_bits() >> (usize::BITS - log_n);
-        if j > i {
-            v.as_mut().swap(i, j);
-        }
-    }
-    v
-}
-
 #[cfg(test)]
 mod tests {
     use super::repeat_value;
-    use crate::core::poly::utils::bit_reverse;
 
     #[test]
     fn repeat_value_0_times_works() {
@@ -89,19 +70,5 @@ mod tests {
     #[test]
     fn repeat_value_3_times_works() {
         assert_eq!(repeat_value(&[1, 2], 3), vec![1, 1, 1, 2, 2, 2]);
-    }
-
-    #[test]
-    fn bit_reverse_works() {
-        assert_eq!(
-            bit_reverse([0, 1, 2, 3, 4, 5, 6, 7]),
-            [0, 4, 2, 6, 1, 5, 3, 7]
-        );
-    }
-
-    #[test]
-    #[should_panic]
-    fn bit_reverse_non_power_of_two_size_fails() {
-        bit_reverse([0, 1, 2, 3, 4, 5]);
     }
 }
